requirements.txt
astroid==2.1.0
attrs==18.2.0
isort==4.3.4
jsonschema==2.6.0
lazy-object-proxy==1.3.1
mccabe==0.6.1
openapi-core==0.7.1
openapi-spec-validator==0.2.4
pylint==2.2.2
PyYAML==3.13
six==1.12.0
strict-rfc3339==0.7
typed-ast==1.2.0
Werkzeug==0.14.1
wrapt==1.11.1
